+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project are documented below.
 
 The format is based on [keep a changelog](http://keepachangelog.com) and this project uses [semantic versioning](http://semver.org).
 
+## [Unreleased]
+
+
 ## [2.2.0] - 2018-11-11
 ### Added
 - New runtime function to send raw realtime envelope data through streams.