+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project are documented below.
 The format is based on [keep a changelog](http://keepachangelog.com) and this project uses [semantic versioning](http://semver.org).
 
 ## [Unreleased]
+### Added
+- New duration field in the tournament API. 
+
 ### Fixed
 - Set friend state correctly when initially adding friends.
 - Join events on tournaments with an end time set now allow users to submit scores.